polish the code
