+++ src/main/kotlin/apply/application/CheaterDtos.kt
@@ -1,5 +1,8 @@
 package apply.application
 
+import apply.domain.applicant.Applicant
+import apply.domain.cheater.Cheater
+import java.time.LocalDateTime
 import javax.validation.constraints.Email
 import javax.validation.constraints.NotBlank
 
@@ -11,3 +14,15 @@ data class CheaterData(
     @field:NotBlank
     var description: String
 )
+
+data class CheaterResponse(
+    val id: Long,
+    val createdDateTime: LocalDateTime,
+    val applicant: Applicant
+) {
+    constructor(cheater: Cheater, applicant: Applicant) : this(
+        cheater.id,
+        cheater.createdDateTime,
+        applicant
+    )
+}
